+++ src/app/(protected)/dashboard/tarefas/components/columns.tsx
@@ -21,8 +21,8 @@ export type Task = {
   associacao: string | null
   tipoId: string | null
 }
-import { DataTableColumnHeader } from "./data-table-column-header"
-import { DataTableRowActions } from "./data-table-row-actions"
+import { DataTableColumnHeader } from "./DataTableColumnHeader"
+import { DataTableRowActions } from "./DataTableRowActions"
 
 export const columns: ColumnDef<Task>[] = [
   {
